api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[argon2]==1.7.4
psycopg[binary]==3.2.2
redis==5.0.8
httpx==0.27.2
openai==1.47.0
qdrant-client==1.11.1
python-multipart==0.0.9
pytest==8.2.2
boto3==1.34.162
loguru==0.7.2
pydantic[email]
